--- feature/exercise/src/main/res/drawable/bg_record_bottom_sheet.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <solid android:color="@color/white300" />
-    <corners android:topLeftRadius="10dp"
-        android:topRightRadius="10dp"/>
-</shape>
