+++ presentation/src/main/java/com/project200/presentation/view/MenuBottomSheetDialog.kt
@@ -1,20 +1,23 @@
-package com.project200.feature.exercise.form
+package com.project200.presentation.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-import com.project200.undabang.feature.exercise.R
-// strings.xml에 정의된 R을 사용하기 위해 실제 R 클래스를 import 해야 할 수 있습니다.
-// 예: import com.project200.undabang.feature.exercise.R
-import com.project200.undabang.feature.exercise.databinding.BottomSheetExerciseMenuBinding
+import com.project200.undabang.presentation.R
+import com.project200.undabang.presentation.databinding.BottomSheetDialogMenuBinding
 
-class ExerciseMenuBottomSheet(
+/** 메뉴 바텀 시트 다이얼로그
+ * @param onEditClicked 수정 버튼 클릭 시 호출되는 콜백
+ * @param onDeleteClicked 삭제 버튼 클릭 시 호출되는 콜백
+ */
+class MenuBottomSheetDialog(
     val onEditClicked: () -> Unit,
-    val onDeleteClicked: () -> Unit
+    val onDeleteClicked: () -> Unit,
+    private val isEditVisible: Boolean = true, // TODO: 타이머 수정 기능이 추가되면 제거 예정
 ) : BottomSheetDialogFragment() {
-    private var _binding: BottomSheetExerciseMenuBinding? = null
+    private var _binding: BottomSheetDialogMenuBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -26,24 +29,26 @@ class ExerciseMenuBottomSheet(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = BottomSheetExerciseMenuBinding.inflate(inflater, container, false)
+        _binding = BottomSheetDialogMenuBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        binding.recordEditBtn.setOnClickListener {
+        binding.editBtn.setOnClickListener {
             onEditClicked()
             dismiss()
         }
 
-        binding.recordDeleteBtn.setOnClickListener {
+        binding.deleteBtn.setOnClickListener {
             onDeleteClicked()
             dismiss()
         }
 
         binding.closeBtn.setOnClickListener { dismiss() }
+
+        binding.editBtn.visibility = if (isEditVisible) View.VISIBLE else View.GONE // TODO: 타이머 수정 기능이 추가되면 제거 예정
     }
 
     override fun onDestroyView() {
