+++ presentation/src/main/res/drawable/ic_edit.xml
@@ -1,0 +1,20 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="20dp"
+    android:viewportWidth="20"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M13.729,4.507L15.494,6.271M14.864,2.952L10.091,7.725C9.844,7.971 9.676,8.285 9.608,8.627L9.167,10.833L11.374,10.392C11.715,10.323 12.029,10.156 12.275,9.909L17.048,5.137C17.191,4.993 17.305,4.823 17.383,4.636C17.46,4.448 17.5,4.247 17.5,4.045C17.5,3.842 17.46,3.641 17.383,3.454C17.305,3.266 17.191,3.096 17.048,2.952C16.904,2.809 16.734,2.695 16.547,2.618C16.359,2.54 16.159,2.5 15.956,2.5C15.753,2.5 15.552,2.54 15.365,2.618C15.177,2.695 15.007,2.809 14.864,2.952Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#212121"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M15.833,12.5V15C15.833,15.442 15.658,15.866 15.345,16.178C15.033,16.491 14.609,16.667 14.167,16.667H5C4.558,16.667 4.134,16.491 3.822,16.178C3.509,15.866 3.333,15.442 3.333,15V5.833C3.333,5.391 3.509,4.967 3.822,4.655C4.134,4.342 4.558,4.167 5,4.167H7.5"
+      android:strokeLineJoin="round"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#212121"
+      android:strokeLineCap="round"/>
+</vector>
