+++ presentation/src/main/res/drawable/ic_trash.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="20dp"
+    android:viewportWidth="20"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M3.333,5.833H16.667M8.333,9.167V14.167M11.667,9.167V14.167M4.167,5.833L5,15.833C5,16.275 5.176,16.699 5.488,17.012C5.801,17.324 6.225,17.5 6.667,17.5H13.333C13.776,17.5 14.199,17.324 14.512,17.012C14.825,16.699 15,16.275 15,15.833L15.833,5.833M7.5,5.833V3.333C7.5,3.112 7.588,2.9 7.744,2.744C7.901,2.588 8.112,2.5 8.333,2.5H11.667C11.888,2.5 12.1,2.588 12.256,2.744C12.412,2.9 12.5,3.112 12.5,3.333V5.833"
+      android:strokeLineJoin="round"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="@color/error_led"
+      android:strokeLineCap="round"/>
+</vector>
