+++ presentation/src/main/res/layout/bottom_sheet_dialog_menu.xml
@@ -6,7 +6,7 @@
     android:orientation="vertical">
 
     <LinearLayout
-        android:id="@+id/record_edit_btn"
+        android:id="@+id/edit_btn"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="horizontal"
@@ -16,7 +16,7 @@
         <ImageView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:src="@drawable/ic_record_edit"
+            android:src="@drawable/ic_edit"
             android:layout_gravity="center" />
         <TextView
             style="@style/content_regular"
@@ -25,11 +25,11 @@
             android:layout_gravity="center"
             android:layout_marginStart="10dp"
             app:layout_constraintTop_toTopOf="parent"
-            android:text="@string/exercise_record_edit" />
+            android:text="@string/bottom_sheet_edit" />
     </LinearLayout>
 
     <LinearLayout
-        android:id="@+id/record_delete_btn"
+        android:id="@+id/delete_btn"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="horizontal"
@@ -49,7 +49,7 @@
             android:layout_marginStart="10dp"
             android:textColor="@color/error_led"
             app:layout_constraintTop_toTopOf="parent"
-            android:text="@string/exercise_record_delete" />
+            android:text="@string/bottom_sheet_delete" />
     </LinearLayout>
 
     <View
