+++ presentation/src/main/res/values/strings.xml
@@ -21,4 +21,7 @@
     <!-- 이용약관, 개인정보처리방침 링크 -->
     <string name="terms_link">https://www.undabang.store/legal/terms-of-service.html</string>
     <string name="privacy_link">https://www.undabang.store/legal/privacy-policy.html</string>
+
+    <string name="bottom_sheet_delete">삭제하기</string>
+    <string name="bottom_sheet_edit">수정하기</string>
 </resources>
